+++ NEWS.md
@@ -1,3 +1,5 @@
 # InquisitR 0.1.0
 
-* Initial CRAN submission.
+* Initial release with basic functionality including:
+  * `boxplotR` function for generating boxplots.
+  * `distributionR` function for creating distribution plots.